removing BitString and OctetString and use BigInteger instead

+++ ecc/src/eccdh/adt/PrivateKey.java
@@ -4,6 +4,8 @@
  */
 package eccdh.adt;
 
+import java.math.BigInteger;
+
 /**
  *
  * @author rolf
@@ -12,5 +14,5 @@ public class PrivateKey {
     /**
      * Der Private Key zu einer elliptischen Kurve und einem gemeinsamen Punkt. 
      */
-    private long d;
+    private BigInteger d;
 }
